implemeted API add/sync/list capability center
